+++ www/go/core/jmap/State.php
@@ -115,7 +115,7 @@ class State extends AbstractState {
 	{
 		$token = $this->getToken();
 		$token->userId = $userId;
-		$success = $token->setAuthenticated();
+		$success = $token->setAuthenticated(false);
 
 		go()->getCache()->delete('token-' . $token->accessToken);
 		go()->getCache()->delete('session-' . $token->accessToken);