+++ CHANGELOG.md
@@ -1,8 +1,13 @@
-6.6.124
+- Core: Switch user will not update last login and login count.
+
+31-10-2022 6.6.124
 - Core: Add extra CSS options to custom comboboxes
 - Billing: add profit and profit percentage fields, also add both fields to customer report
 - Billing: add customer report export
 - Comments: Only the creator or admin may edit and delete comments
+- Core: Added "Is empty" filter to custom date filter
+- Calendar: When an appointment was opened from an email it was set as declined for the organizer immediately
+- Email: Fix: open linked email in popup
 
 24-10-2022: 6.6.123
 - Core: Remember user selected language